pyproject.toml
[project]
name = "Personal-Site"
version = "0.1.0"
description = "Personal Site of Stephen Steyaert"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=4.2.23",
    "django-hosts>=7.0.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "freezegun>=1.5.3",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
    "pytest-factoryboy>=2.8.1",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.1.1",
]

prod = [
    "gunicorn>=23.0.0",
]
